Update user profile UI
