+++ app/src/main/java/nl/mwsoft/www/superheromatch/viewLayer/main/adapter/UserProfileSettingsViewPagerAdapter.java
@@ -5,9 +5,8 @@ import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
 import androidx.fragment.app.FragmentPagerAdapter;
 
-import nl.mwsoft.www.superheromatch.viewLayer.main.fragment.UserProfileSettingsDiscoverFragment;
-import nl.mwsoft.www.superheromatch.viewLayer.main.fragment.UserProfileSettingsNotificationsFragment;
-import nl.mwsoft.www.superheromatch.viewLayer.main.fragment.UserProfileSettingsSuggestionsFragment;
+import nl.mwsoft.www.superheromatch.viewLayer.main.fragment.profile.UserProfileSettingsNotificationsFragment;
+import nl.mwsoft.www.superheromatch.viewLayer.main.fragment.profile.UserProfileSettingsSuggestionsFragment;
 
 public class UserProfileSettingsViewPagerAdapter extends FragmentPagerAdapter {
