+++ app/src/main/java/nl/mwsoft/www/superheromatch/viewLayer/main/fragment/profile/ProfileDetailsFragment.java
@@ -1,44 +1,48 @@
-package nl.mwsoft.www.superheromatch.viewLayer.main.fragment;
+package nl.mwsoft.www.superheromatch.viewLayer.main.fragment.profile;
 
 import android.content.Context;
+import android.graphics.Typeface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ImageView;
+import android.widget.TextView;
 import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
-import com.squareup.picasso.Picasso;
-
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.Unbinder;
 import nl.mwsoft.www.superheromatch.R;
+import nl.mwsoft.www.superheromatch.modelLayer.model.User;
 import nl.mwsoft.www.superheromatch.viewLayer.main.activity.MainActivity;
 
-public class ImageDetailFragment extends Fragment {
+public class ProfileDetailsFragment extends Fragment {
 
-    @BindView(R.id.ivImageDetail)
-    ImageView ivImageDetail;
+    @BindView(R.id.tvUserProfileDetailsSuperPower)
+    TextView tvUserProfileDetailsAbout;
+    @BindView(R.id.tvUserProfileDetailsSuperPowerDesc)
+    TextView tvUserProfileDetailsAboutDesc;
+
     private Unbinder unbinder;
     private MainActivity mainActivity;
-    private String imageUrl;
-    private static final String IMAGE_URL = "image_url";
+    private User user;
+    private final static String USER = "user";
 
-    public static ImageDetailFragment newInstance(String imageUrl) {
+    public static ProfileDetailsFragment newInstance(User user) {
         Bundle args = new Bundle();
-        args.putString(IMAGE_URL, imageUrl);
-        ImageDetailFragment fragment = new ImageDetailFragment();
+        args.putParcelable(USER, user);
+        ProfileDetailsFragment fragment = new ProfileDetailsFragment();
         fragment.setArguments(args);
         return fragment;
     }
 
+
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_image_detail, container, false);
+        View view = inflater.inflate(R.layout.fragment_user_profile_details, container, false);
         unbinder = ButterKnife.bind(this, view);
         return view;
     }
@@ -46,24 +50,23 @@ public class ImageDetailFragment extends Fragment {
     @Override
     public void onViewCreated(View view, @Nullable Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
-        mainActivity.setToolbarStatusBarColorBlack();
         Bundle arguments = getArguments();
         if(arguments != null){
-            imageUrl = arguments.getString(IMAGE_URL);
-            if(imageUrl == null){
-                Toast.makeText(mainActivity,"Something went wrong!",Toast.LENGTH_SHORT).show();
-            }else {
-                Picasso.with(mainActivity).load(R.drawable.user_512).into(ivImageDetail);
-            }
+            user = arguments.getParcelable(USER);
+            loadUserProfileDetails(user);
         }else{
-            Toast.makeText(mainActivity,"Something went wrong!",Toast.LENGTH_SHORT).show();
+            Toast.makeText(mainActivity, "Something went wrong!", Toast.LENGTH_LONG).show();
         }
+    }
+
+    private void loadUserProfileDetails(User user) {
+        tvUserProfileDetailsAbout.setTypeface(tvUserProfileDetailsAbout.getTypeface(), Typeface.BOLD);
+        tvUserProfileDetailsAboutDesc.setText(user.getSuperPower());
     }
 
     @Override
     public void onDestroy() {
         super.onDestroy();
-        mainActivity.setToolbarStatusBarColorPrimary();
         unbinder.unbind();
     }
