+++ app/src/main/java/nl/mwsoft/www/superheromatch/viewLayer/main/fragment/profile/UserProfileImageFragment.java
@@ -1,19 +1,16 @@
-package nl.mwsoft.www.superheromatch.viewLayer.main.fragment;
+package nl.mwsoft.www.superheromatch.viewLayer.main.fragment.profile;
 
 import android.content.Context;
 import android.graphics.Typeface;
 import android.os.Bundle;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ImageView;
 import android.widget.TextView;
-import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.bumptech.glide.Glide;
 
@@ -23,16 +20,16 @@ import butterknife.Unbinder;
 import nl.mwsoft.www.superheromatch.R;
 import nl.mwsoft.www.superheromatch.viewLayer.main.activity.MainActivity;
 
-public class SuggestionProfileImageFragment extends Fragment {
+public class UserProfileImageFragment  extends Fragment {
 
-    @BindView(R.id.ivSuggestionProfilePic)
-    ImageView ivSuggestionProfilePic;
-    @BindView(R.id.tvSuggestionNameAge)
-    TextView tvSuggestionNameAge;
-    @BindView(R.id.tvSuggestionCity)
-    TextView tvSuggestionCity;
-    @BindView(R.id.tvSuggestionSuperPowerDesc)
-    TextView tvSuggestionSuperPowerDesc;
+    @BindView(R.id.ivUserProfilePic)
+    ImageView ivUserProfilePic;
+    @BindView(R.id.tvUserNameAge)
+    TextView tvUserNameAge;
+    @BindView(R.id.tvUserCity)
+    TextView tvUserCity;
+    @BindView(R.id.tvUserSuperPowerDesc)
+    TextView tvUserSuperPowerDesc;
     private Unbinder unbinder;
     private MainActivity mainActivity;
     private static final String PIC_URL = "picUrl";
@@ -46,14 +43,14 @@ public class SuggestionProfileImageFragment extends Fragment {
     private String city;
     private String superpower;
 
-    public static SuggestionProfileImageFragment newInstance(String picUrl, String name, int age, String city, String superpower) {
+    public static UserProfileImageFragment newInstance(String picUrl, String name, int age, String city, String superpower) {
         Bundle args = new Bundle();
         args.putString(PIC_URL, picUrl);
         args.putString(NAME, name);
         args.putInt(AGE, age);
         args.putString(CITY, city);
         args.putString(SUPERPOWER, superpower);
-        SuggestionProfileImageFragment fragment = new SuggestionProfileImageFragment();
+        UserProfileImageFragment fragment = new UserProfileImageFragment();
         fragment.setArguments(args);
 
         return fragment;
@@ -61,7 +58,7 @@ public class SuggestionProfileImageFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_suggestion_profile_image, container, false);
+        View view = inflater.inflate(R.layout.fragment_user_profile_image, container, false);
         unbinder = ButterKnife.bind(this, view);
         return view;
     }
@@ -84,83 +81,79 @@ public class SuggestionProfileImageFragment extends Fragment {
             return;
         }
 
-        ivSuggestionProfilePic.setClipToOutline(true);
+        ivUserProfilePic.setClipToOutline(true);
 
         switch (picUrl) {
-            case "test":
-                Glide.with(mainActivity)
-                        .load(R.drawable.test)
-                        .into(ivSuggestionProfilePic);
-                break;
             case "test1":
                 Glide.with(mainActivity)
                         .load(R.drawable.test1)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test2":
                 Glide.with(mainActivity)
                         .load(R.drawable.test2)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test3":
                 Glide.with(mainActivity)
                         .load(R.drawable.test3)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test4":
                 Glide.with(mainActivity)
                         .load(R.drawable.test4)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test5":
                 Glide.with(mainActivity)
                         .load(R.drawable.test5)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test6":
                 Glide.with(mainActivity)
                         .load(R.drawable.test6)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test7":
                 Glide.with(mainActivity)
                         .load(R.drawable.test7)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test8":
                 Glide.with(mainActivity)
                         .load(R.drawable.test8)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test9":
                 Glide.with(mainActivity)
                         .load(R.drawable.test9)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test10":
                 Glide.with(mainActivity)
                         .load(R.drawable.test10)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test11":
                 Glide.with(mainActivity)
                         .load(R.drawable.test11)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
             case "test12":
                 Glide.with(mainActivity)
                         .load(R.drawable.test12)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
                 break;
+            case "test":
             default:
                 Glide.with(mainActivity)
                         .load(R.drawable.test)
-                        .into(ivSuggestionProfilePic);
+                        .into(ivUserProfilePic);
         }
 
-        tvSuggestionNameAge.bringToFront();
-        tvSuggestionNameAge.setTypeface(tvSuggestionNameAge.getTypeface(), Typeface.BOLD);
-        tvSuggestionNameAge.setText(
+        tvUserNameAge.bringToFront();
+        tvUserNameAge.setTypeface(tvUserNameAge.getTypeface(), Typeface.BOLD);
+        tvUserNameAge.setText(
                 getString(
                         R.string.name_age,
                         name,
@@ -168,10 +161,10 @@ public class SuggestionProfileImageFragment extends Fragment {
                 )
         );
 
-        tvSuggestionCity.setTypeface(tvSuggestionCity.getTypeface(), Typeface.BOLD);
-        tvSuggestionCity.setText(city);
+        tvUserCity.setTypeface(tvUserCity.getTypeface(), Typeface.BOLD);
+        tvUserCity.setText(city);
 
-        tvSuggestionSuperPowerDesc.setText(superpower);
+        tvUserSuperPowerDesc.setText(superpower);
     }
 
     @Override
@@ -185,5 +178,4 @@ public class SuggestionProfileImageFragment extends Fragment {
         super.onAttach(context);
         mainActivity = (MainActivity) context;
     }
-
 }
