+++ app/src/main/java/nl/mwsoft/www/superheromatch/viewLayer/main/fragment/suggestions/SuggestionFragment.java
@@ -1,53 +1,49 @@
-package nl.mwsoft.www.superheromatch.viewLayer.main.fragment;
+package nl.mwsoft.www.superheromatch.viewLayer.main.fragment.suggestions;
 
 import android.content.Context;
-import android.graphics.Typeface;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
+import androidx.viewpager.widget.ViewPager;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Toast;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
-import butterknife.OnClick;
 import butterknife.Unbinder;
 import nl.mwsoft.www.superheromatch.R;
 import nl.mwsoft.www.superheromatch.modelLayer.model.Superhero;
 import nl.mwsoft.www.superheromatch.viewLayer.main.activity.MainActivity;
+import nl.mwsoft.www.superheromatch.viewLayer.main.adapter.SuggestionProfileViewPagerAdapter;
 
-public class SuggestionDescriptionFragment extends Fragment {
+public class SuggestionFragment extends Fragment {
 
-    @BindView(R.id.ivSuggestionDescriptionClose)
-    ImageView ivSuggestionDescriptionClose;
-    @BindView(R.id.tvSuggestionDescriptionNameAge)
-    TextView tvSuggestionDescriptionNameAge;
-    @BindView(R.id.tvSuggestionDescriptionCity)
-    TextView tvSuggestionDescriptionCity;
-    @BindView(R.id.tvSuggestionDescriptionSuperpower)
-    TextView tvSuggestionDescriptionSuperpower;
+    @BindView(R.id.vpSuggestionUserProfile)
+    ViewPager vpSuggestionUserProfile;
+    private SuggestionProfileViewPagerAdapter suggestionUserProfileViewPagerAdapter;
     private Unbinder unbinder;
     private MainActivity mainActivity;
     private Superhero superhero;
     private static final String SUPERHERO = "superhero";
 
-    public static SuggestionDescriptionFragment newInstance(Superhero superhero) {
+    public static SuggestionFragment newInstance(Superhero superhero) {
         Bundle args = new Bundle();
         args.putParcelable(SUPERHERO, superhero);
-        SuggestionDescriptionFragment fragment = new SuggestionDescriptionFragment();
+        SuggestionFragment fragment = new SuggestionFragment();
         fragment.setArguments(args);
+
         return fragment;
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_suggestion_description, container, false);
+        View view = inflater.inflate(R.layout.fragment_suggestion, container, false);
         unbinder = ButterKnife.bind(this, view);
+
         return view;
     }
 
@@ -58,21 +54,12 @@ public class SuggestionDescriptionFragment extends Fragment {
         Bundle arguments = getArguments();
         if (arguments != null) {
             superhero = arguments.getParcelable(SUPERHERO);
+
             if (superhero == null) {
                 superhero = new Superhero();
                 Toast.makeText(mainActivity, "Something went wrong!", Toast.LENGTH_SHORT).show();
             } else {
-                tvSuggestionDescriptionNameAge.setTypeface(tvSuggestionDescriptionNameAge.getTypeface(), Typeface.BOLD);
-                tvSuggestionDescriptionNameAge.setText(
-                        getString(
-                                R.string.name_age,
-                                superhero.getSuperheroName(),
-                                superhero.getAge()
-                        )
-                );
-
-                tvSuggestionDescriptionCity.setText(superhero.getCity());
-                tvSuggestionDescriptionSuperpower.setText(superhero.getSuperpower());
+                configureViewPager(superhero);
             }
         } else {
             superhero = new Superhero();
@@ -92,8 +79,13 @@ public class SuggestionDescriptionFragment extends Fragment {
         mainActivity = (MainActivity) context;
     }
 
-    @OnClick(R.id.ivSuggestionDescriptionClose)
-    public void onSuggestionDescriptionClose() {
-        mainActivity.closeSuggestionDescriptionWindow();
+    private void configureViewPager(Superhero suggestion) {
+        suggestionUserProfileViewPagerAdapter = new SuggestionProfileViewPagerAdapter(
+                getChildFragmentManager(),
+                suggestion
+        );
+
+        vpSuggestionUserProfile.setOffscreenPageLimit(1);
+        vpSuggestionUserProfile.setAdapter(suggestionUserProfileViewPagerAdapter);
     }
 }
