+++ app/src/main/res/layout/fragment_suggestion.xml
@@ -5,128 +5,10 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/colorWhite">
 
-<!--    <ImageView-->
-<!--        android:id="@+id/ivSuggestionProfilePic"-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:layout_marginStart="5dp"-->
-<!--        android:layout_marginTop="5dp"-->
-<!--        android:layout_marginEnd="5dp"-->
-<!--        android:background="@drawable/my_image_view"-->
-<!--        android:clickable="true"-->
-<!--        android:contentDescription="@string/user_profile_image"-->
-<!--        android:elevation="5dp"-->
-<!--        android:focusable="true"-->
-<!--        android:maxHeight="435dp"-->
-<!--        android:src="@drawable/test" />-->
-
-<!--    <RelativeLayout-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="125dp"-->
-<!--        android:layout_alignParentBottom="true"-->
-<!--        android:layout_marginStart="20dp"-->
-<!--        android:layout_marginEnd="20dp"-->
-<!--        android:layout_marginBottom="125dp"-->
-<!--        android:background="@drawable/black_gradient"-->
-<!--        android:elevation="10dp">-->
-
-<!--        <TextView-->
-<!--            android:id="@+id/tvSuggestionNameAge"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_above="@+id/tvSuggestionCity"-->
-<!--            android:layout_centerHorizontal="true"-->
-<!--            android:layout_marginStart="20dp"-->
-<!--            android:elevation="10dp"-->
-<!--            android:shadowColor="@color/colorDarkGrey"-->
-<!--            android:shadowDx="2"-->
-<!--            android:shadowDy="2"-->
-<!--            android:shadowRadius="2"-->
-<!--            android:text="Superhero, 34"-->
-<!--            android:textColor="@color/colorWhite"-->
-<!--            android:textSize="24sp"-->
-<!--            android:textStyle="bold" />-->
-
-<!--        <TextView-->
-<!--            android:id="@+id/tvSuggestionCity"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_above="@+id/tvSuggestionSuperPowerDesc"-->
-<!--            android:layout_centerHorizontal="true"-->
-<!--            android:layout_marginStart="20dp"-->
-<!--            android:elevation="10dp"-->
-<!--            android:shadowColor="@color/colorDarkGrey"-->
-<!--            android:shadowDx="2"-->
-<!--            android:shadowDy="2"-->
-<!--            android:shadowRadius="2"-->
-<!--            android:text="Utrecht"-->
-<!--            android:textColor="@color/colorWhite"-->
-<!--            android:textSize="18sp"-->
-<!--            android:textStyle="bold" />-->
-
-<!--        <TextView-->
-<!--            android:id="@+id/tvSuggestionSuperPowerDesc"-->
-<!--            android:layout_width="wrap_content"-->
-<!--            android:layout_height="wrap_content"-->
-<!--            android:layout_alignParentBottom="true"-->
-<!--            android:layout_centerHorizontal="true"-->
-<!--            android:layout_marginStart="20dp"-->
-<!--            android:layout_marginEnd="20dp"-->
-<!--            android:layout_marginBottom="12dp"-->
-<!--            android:elevation="10dp"-->
-<!--            android:shadowColor="@color/colorDarkGrey"-->
-<!--            android:shadowDx="-1"-->
-<!--            android:shadowDy="-1"-->
-<!--            android:shadowRadius="1"-->
-<!--            android:text="@string/test_superpower"-->
-<!--            android:textColor="@color/colorWhite"-->
-<!--            android:textSize="14sp" />-->
-<!--    </RelativeLayout>-->
-
     <androidx.viewpager.widget.ViewPager
         android:id="@+id/vpSuggestionUserProfile"
         android:layout_width="match_parent"
         android:layout_height="435dp"
         app:layout_behavior="@string/appbar_scrolling_view_behavior" />
-
-<!--    <ImageView-->
-<!--        android:id="@+id/ivSuggestionDislike"-->
-<!--        android:layout_width="45dp"-->
-<!--        android:layout_height="45dp"-->
-<!--        android:layout_alignParentStart="true"-->
-<!--        android:layout_alignParentBottom="true"-->
-<!--        android:layout_marginStart="25dp"-->
-<!--        android:layout_marginBottom="65dp"-->
-<!--        android:clickable="true"-->
-<!--        android:contentDescription="@string/dislike"-->
-<!--        android:focusable="true"-->
-<!--        android:src="@drawable/dislike_512" />-->
-
-<!--    <ImageView-->
-<!--        android:id="@+id/ivSuperPowerIconSuggestion"-->
-<!--        android:layout_width="45dp"-->
-<!--        android:layout_height="45dp"-->
-<!--        android:layout_alignParentBottom="true"-->
-<!--        android:layout_centerHorizontal="true"-->
-<!--        android:layout_marginBottom="65dp"-->
-<!--        android:clickable="true"-->
-<!--        android:contentDescription="@string/super_power"-->
-<!--        android:focusable="true"-->
-<!--        android:src="@drawable/superpower_256" />-->
-
-<!--    <ImageView-->
-<!--        android:id="@+id/ivSuggestionLike"-->
-<!--        android:layout_width="45dp"-->
-<!--        android:layout_height="45dp"-->
-<!--        android:layout_alignParentEnd="true"-->
-<!--        android:layout_alignParentBottom="true"-->
-<!--        android:layout_marginTop="25dp"-->
-<!--        android:layout_marginEnd="25dp"-->
-<!--        android:layout_marginBottom="65dp"-->
-<!--        android:clickable="true"-->
-<!--        android:contentDescription="@string/like"-->
-<!--        android:focusable="true"-->
-<!--        android:src="@drawable/like_512"-->
-<!--        android:visibility="visible" />-->
 
 </RelativeLayout>
